requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.36.0
crispy-bootstrap5==0.7
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.11
django-allauth==0.57.2
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-summernote==0.8.20.0
gunicorn==20.1.0
idna==3.6
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.18
webencodings==0.5.1
whitenoise==5.3.0
